requirements.txt
sqlalchemy==1.4.50
psycopg2==2.9.9
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
aiohttp==3.9.1
pika==1.3.1
redis==5.0.1